+++ examples/models/ExampleModel.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace unclead\widgets\examples\models;
+namespace unclead\multipleinput\examples\models;
 
 use yii\base\Model;
 use yii\helpers\ArrayHelper;
@@ -10,7 +10,7 @@ use yii\validators\RequiredValidator;
 
 /**
  * Class ExampleModel
- * @package unclead\widgets\examples\actions
+ * @package unclead\multipleinput\examples\actions
  */
 class ExampleModel extends Model
 {
@@ -93,6 +93,7 @@ class ExampleModel extends Model
     {
         return [
             ['title', 'required'],
+            ['title', 'string', 'min' => 5],
             ['emails', 'validateEmails'],
             ['phones', 'validatePhones'],
             ['schedule', 'validateSchedule'],